Extract chat history helpers
