+++ service-worker.js
@@ -74,6 +74,7 @@ const APP_SHELL = [
   '/js/chat-attestation.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
+  '/js/chat-history.js',
   '/js/chat-continuation.js',
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
