+++ tests/test-correctness-phase2.js
@@ -130,6 +130,7 @@ const pwaAppShellAssets = [
   '/js/chat-attestation.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
+  '/js/chat-history.js',
   '/js/chat-continuation.js',
   '/js/chat-prompt-context.js',
   '/js/chat-summaries.js',
